Added version 5.5.2805

+++ chromium/ash/services/cellular_setup/euicc.cc
@@ -15,14 +15,14 @@
 #include "base/metrics/histogram_macros.h"
 #include "base/strings/strcat.h"
 #include "base/time/time.h"
-#include "chromeos/network/cellular_connection_handler.h"
-#include "chromeos/network/cellular_esim_installer.h"
-#include "chromeos/network/cellular_esim_profile.h"
-#include "chromeos/network/cellular_inhibitor.h"
-#include "chromeos/network/hermes_metrics_util.h"
-#include "chromeos/network/network_connection_handler.h"
-#include "chromeos/network/network_event_log.h"
-#include "chromeos/network/network_state_handler.h"
+#include "chromeos/ash/components/network/cellular_connection_handler.h"
+#include "chromeos/ash/components/network/cellular_esim_installer.h"
+#include "chromeos/ash/components/network/cellular_esim_profile.h"
+#include "chromeos/ash/components/network/cellular_inhibitor.h"
+#include "chromeos/ash/components/network/hermes_metrics_util.h"
+#include "chromeos/ash/components/network/network_connection_handler.h"
+#include "chromeos/ash/components/network/network_event_log.h"
+#include "chromeos/ash/components/network/network_state_handler.h"
 #include "components/device_event_log/device_event_log.h"
 #include "components/qr_code_generator/qr_code_generator.h"
 #include "dbus/object_path.h"
@@ -281,8 +281,7 @@ void Euicc::OnRequestPendingProfilesResult(
   mojom::ESimOperationResult operation_result;
 
   if (status != HermesResponseStatus::kSuccess) {
-    NET_LOG(ERROR) << "Request Pending events failed status="
-                   << static_cast<int>(status);
+    NET_LOG(ERROR) << "Request Pending events failed status=" << status;
     metrics_result = RequestPendingProfilesResult::kHermesRequestFailed;
     operation_result = mojom::ESimOperationResult::kFailure;
   } else {
