+++ chromium/base/system/sys_info_mac.mm
@@ -16,6 +16,7 @@
 #include "base/mac/mac_util.h"
 #include "base/mac/scoped_mach_port.h"
 #include "base/notreached.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/process/process_metrics.h"
 #include "base/strings/string_util.h"
 #include "base/strings/stringprintf.h"
@@ -57,9 +58,9 @@ void SysInfo::OperatingSystemVersionNumbers(int32_t* major_version,
                                             int32_t* bugfix_version) {
   NSOperatingSystemVersion version =
       [[NSProcessInfo processInfo] operatingSystemVersion];
-  *major_version = version.majorVersion;
-  *minor_version = version.minorVersion;
-  *bugfix_version = version.patchVersion;
+  *major_version = static_cast<int32_t>(version.majorVersion);
+  *minor_version = static_cast<int32_t>(version.minorVersion);
+  *bugfix_version = static_cast<int32_t>(version.patchVersion);
 }
 
 // static
@@ -74,7 +75,7 @@ std::string SysInfo::OperatingSystemArchitecture() {
 }
 
 // static
-int64_t SysInfo::AmountOfPhysicalMemoryImpl() {
+uint64_t SysInfo::AmountOfPhysicalMemoryImpl() {
   struct host_basic_info hostinfo;
   mach_msg_type_number_t count = HOST_BASIC_INFO_COUNT;
   base::mac::ScopedMachSendRight host(mach_host_self());
@@ -85,17 +86,17 @@ int64_t SysInfo::AmountOfPhysicalMemoryImpl() {
     return 0;
   }
   DCHECK_EQ(HOST_BASIC_INFO_COUNT, count);
-  return static_cast<int64_t>(hostinfo.max_mem);
+  return hostinfo.max_mem;
 }
 
 // static
-int64_t SysInfo::AmountOfAvailablePhysicalMemoryImpl() {
+uint64_t SysInfo::AmountOfAvailablePhysicalMemoryImpl() {
   SystemMemoryInfoKB info;
   if (!GetSystemMemoryInfo(&info))
     return 0;
   // We should add inactive file-backed memory also but there is no such
   // information from Mac OS unfortunately.
-  return static_cast<int64_t>(info.free + info.speculative) * 1024;
+  return checked_cast<uint64_t>(info.free + info.speculative) * 1024;
 }
 
 // static
