+++ chromium/ui/views/controls/editable_combobox/editable_combobox.cc
@@ -156,12 +156,11 @@ class EditableCombobox::EditableComboboxMenuModel
       return;
     items_shown_.clear();
     if (show_on_empty_ || !owner_->GetText().empty()) {
-      for (int i = 0; i < combobox_model_->GetItemCount(); ++i) {
+      for (size_t i = 0; i < combobox_model_->GetItemCount(); ++i) {
         if (!filter_on_edit_ ||
             base::StartsWith(combobox_model_->GetItemAt(i), owner_->GetText(),
                              base::CompareCase::INSENSITIVE_ASCII)) {
-          items_shown_.push_back(
-              {static_cast<size_t>(i), combobox_model_->IsItemEnabledAt(i)});
+          items_shown_.push_back({i, combobox_model_->IsItemEnabledAt(i)});
         }
       }
     }
@@ -177,8 +176,8 @@ class EditableCombobox::EditableComboboxMenuModel
     return MenuConfig::instance().check_selected_combobox_item;
   }
 
-  std::u16string GetItemTextAt(int index, bool showing_password_text) const {
-    int index_in_model = items_shown_[index].index;
+  std::u16string GetItemTextAt(size_t index, bool showing_password_text) const {
+    size_t index_in_model = items_shown_[index].index;
     std::u16string text = combobox_model_->GetItemAt(index_in_model);
     return showing_password_text
                ? text
@@ -186,7 +185,7 @@ class EditableCombobox::EditableComboboxMenuModel
                                 gfx::RenderText::kPasswordReplacementChar);
   }
 
-  ui::ImageModel GetIconAt(int index) const override {
+  ui::ImageModel GetIconAt(size_t index) const override {
     return combobox_model_->GetDropDownIconAt(items_shown_[index].index);
   }
 
@@ -199,7 +198,7 @@ class EditableCombobox::EditableComboboxMenuModel
     observation_.Reset();
   }
 
-  int GetItemCount() const override { return items_shown_.size(); }
+  size_t GetItemCount() const override { return items_shown_.size(); }
 
  private:
   struct ShownItem {
@@ -207,62 +206,62 @@ class EditableCombobox::EditableComboboxMenuModel
     bool enabled;
   };
   bool HasIcons() const override {
-    for (int i = 0; i < GetItemCount(); ++i) {
+    for (size_t i = 0; i < GetItemCount(); ++i) {
       if (!GetIconAt(i).IsEmpty())
         return true;
     }
     return false;
   }
 
-  ItemType GetTypeAt(int index) const override {
+  ItemType GetTypeAt(size_t index) const override {
     return UseCheckmarks() ? TYPE_CHECK : TYPE_COMMAND;
   }
 
-  ui::MenuSeparatorType GetSeparatorTypeAt(int index) const override {
+  ui::MenuSeparatorType GetSeparatorTypeAt(size_t index) const override {
     return ui::NORMAL_SEPARATOR;
   }
 
-  int GetCommandIdAt(int index) const override {
+  int GetCommandIdAt(size_t index) const override {
     constexpr int kFirstMenuItemId = 1000;
-    return index + kFirstMenuItemId;
+    return static_cast<int>(index) + kFirstMenuItemId;
   }
 
-  std::u16string GetLabelAt(int index) const override {
+  std::u16string GetLabelAt(size_t index) const override {
     std::u16string text = GetItemTextAt(index, owner_->showing_password_text_);
     base::i18n::AdjustStringForLocaleDirection(&text);
     return text;
   }
 
-  bool IsItemDynamicAt(int index) const override { return false; }
+  bool IsItemDynamicAt(size_t index) const override { return false; }
 
-  const gfx::FontList* GetLabelFontListAt(int index) const override {
+  const gfx::FontList* GetLabelFontListAt(size_t index) const override {
     return &owner_->GetFontList();
   }
 
-  bool GetAcceleratorAt(int index,
+  bool GetAcceleratorAt(size_t index,
                         ui::Accelerator* accelerator) const override {
     return false;
   }
 
-  bool IsItemCheckedAt(int index) const override {
+  bool IsItemCheckedAt(size_t index) const override {
     return UseCheckmarks() &&
            combobox_model_->GetItemAt(items_shown_[index].index) ==
                owner_->GetText();
   }
 
-  int GetGroupIdAt(int index) const override { return -1; }
+  int GetGroupIdAt(size_t index) const override { return -1; }
 
-  ui::ButtonMenuItemModel* GetButtonMenuItemAt(int index) const override {
+  ui::ButtonMenuItemModel* GetButtonMenuItemAt(size_t index) const override {
     return nullptr;
   }
 
-  bool IsEnabledAt(int index) const override {
+  bool IsEnabledAt(size_t index) const override {
     return items_shown_[index].enabled;
   }
 
-  void ActivatedAt(int index) override { owner_->OnItemSelected(index); }
+  void ActivatedAt(size_t index) override { owner_->OnItemSelected(index); }
 
-  MenuModel* GetSubmenuModelAt(int index) const override { return nullptr; }
+  MenuModel* GetSubmenuModelAt(size_t index) const override { return nullptr; }
 
   raw_ptr<EditableCombobox> owner_;            // Weak. Owns |this|.
   raw_ptr<ui::ComboboxModel> combobox_model_;  // Weak.
@@ -419,15 +418,15 @@ void EditableCombobox::RevealPasswords(bool revealed) {
   menu_model_->UpdateItemsShown();
 }
 
-int EditableCombobox::GetItemCountForTest() {
+size_t EditableCombobox::GetItemCountForTest() {
   return menu_model_->GetItemCount();
 }
 
-std::u16string EditableCombobox::GetItemForTest(int index) {
+std::u16string EditableCombobox::GetItemForTest(size_t index) {
   return menu_model_->GetItemTextAt(index, showing_password_text_);
 }
 
-ui::ImageModel EditableCombobox::GetIconForTest(int index) {
+ui::ImageModel EditableCombobox::GetIconForTest(size_t index) {
   return menu_model_->GetIconAt(index);
 }
 
@@ -493,7 +492,7 @@ void EditableCombobox::CloseMenu() {
   pre_target_handler_.reset();
 }
 
-void EditableCombobox::OnItemSelected(int index) {
+void EditableCombobox::OnItemSelected(size_t index) {
   // |textfield_| can hide the characters on its own so we read the actual
   // characters instead of gfx::RenderText::kPasswordReplacementChar characters.
   std::u16string selected_item_text =
